third_party/blink/web_tests/external/wpt/html/browsers/history/the-location-interface/location-ancestor-origins-new-object-expected.txt

This is a testharness.js-based test.
[FAIL] location.ancestorOrigins new object semantics
  assert_equals: should return the same object before removing the iframe expected object "[object DOMStringList]" but got object "[object DOMStringList]"
Harness: the test ran to completion.
